# docs/swagger.yaml
basePath: /api
definitions:
...
  response.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: Неверный формат запроса
        type: string
    type: object
...
  restaurant.DishResponse:
    properties:
      description:
        example: Сочный бургер с сыром
        type: string
      id:
        example: "1"
        type: string
      image_url:
        type: string
      name:
        example: Чизбургер
        type: string
      price:
        example: 199000000
        type: integer
    type: object
  restaurant.DishWithBrand:
    properties:
      description:
        type: string
      id:
        type: string
      image_url:
        type: string
      name:
        type: string
      price:
        type: integer
      restaurant_brand_id:
        type: string
    type: object
  restaurant.DishesResponse:
    properties:
      dishes:
        items:
          $ref: '#/definitions/restaurant.DishResponse'
        type: array
    type: object
  restaurant.RestaurantBrandResponse:
    properties:
      description:
        example: Острые крылышки
        type: string
      id:
        example: "1"
        type: string
      logo_url:
        type: string
      name:
        example: KFC
        type: string
      promotion_tier:
        example: 2
        type: integer
    type: object
  restaurant.RestaurantBrandsResponse:
    properties:
      restaurants:
        items:
          $ref: '#/definitions/restaurant.RestaurantBrandResponse'
        type: array
    type: object
  restaurant.SearchAllResponse:
    properties:
      dishes:
        items:
          $ref: '#/definitions/restaurant.DishWithBrand'
        type: array
      restaurants:
        items:
          $ref: '#/definitions/restaurant.RestaurantBrandResponse'
        type: array
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API бэкенда для проекта "Delivery Club" от команды NaNcats
  title: NaNcats Delivery API
  version: "1.0"
paths:
...
  /categories/{slug}/restaurants:
    get:
      consumes:
      - application/json
      description: 'Возвращает список брендов ресторанов, отфильтрованный по переданному
        слагу категории (например: pizza, sushi)'
      parameters:
      - description: Слаг категории
        in: path
        name: slug
        required: true
        type: string
      - description: Количество возвращаемых элементов (по умолчанию 20)
        in: query
        name: limit
        type: integer
      - description: Смещение для пагинации (по умолчанию 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное получение ресторанов по категории
          schema:
            $ref: '#/definitions/restaurant.RestaurantBrandsResponse'
        "400":
          description: Отсутствует слаг категории
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Получение ресторанов по категории
      tags:
      - categories
...
  /restaurants:
    get:
      consumes:
      - application/json
      description: Возвращает список всех брендов ресторанов с поддержкой пагинации
      parameters:
      - description: Количество возвращаемых элементов (по умолчанию 20)
        in: query
        name: limit
        type: integer
      - description: Смещение для пагинации (по умолчанию 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное получение списка ресторанов
          schema:
            $ref: '#/definitions/restaurant.RestaurantBrandsResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Получение списка ресторанов
      tags:
      - restaurants
  /restaurants/{id}:
    get:
      consumes:
      - application/json
      description: Возвращает детальную информацию о бренде ресторана по его уникальному
        идентификатору
      parameters:
      - description: ID ресторана
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешное получение ресторана
          schema:
            $ref: '#/definitions/restaurant.RestaurantBrandResponse'
        "400":
          description: Неверный формат ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Ресторан не найден
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Получение ресторана по ID
      tags:
      - restaurants
  /restaurants/{id}/dishes:
    get:
      consumes:
      - application/json
      description: Возвращает список блюд для конкретного ресторана с возможностью
        пагинации и поиска по названию
      parameters:
      - description: ID ресторана
        in: path
        name: id
        required: true
        type: string
      - description: Количество возвращаемых элементов (по умолчанию 20)
        in: query
        name: limit
        type: integer
      - description: Смещение для пагинации (по умолчанию 0)
        in: query
        name: offset
        type: integer
      - description: Поисковый запрос для фильтрации блюд
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешное получение списка блюд
          schema:
            $ref: '#/definitions/restaurant.DishesResponse'
        "400":
          description: Неверный формат ID ресторана
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Блюда или ресторан не найдены
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Получение списка блюд ресторана
      tags:
      - restaurants
...
  /search:
    get:
      consumes:
      - application/json
      description: Осуществляет поиск по ресторанам и блюдам на основе переданного
        ключевого слова
      parameters:
      - description: Поисковый запрос
        in: query
        name: q
        required: true
        type: string
      - description: Количество возвращаемых элементов (по умолчанию 20)
        in: query
        name: limit
        type: integer
      - description: Смещение для пагинации (по умолчанию 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешный поиск (возвращает списки ресторанов и блюд)
          schema:
            $ref: '#/definitions/restaurant.SearchAllResponse'
        "400":
          description: Отсутствует поисковый запрос
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Глобальный поиск
      tags:
      - search
...
swagger: "2.0"